cleaning on moving post-merge
